requirements.txt
cx_Oracle==8.3.0
Cython==3.0.12
et_xmlfile==2.0.0
numpy==2.2.5
openpyxl==3.1.5
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
tabulate==0.9.0
xlwt==1.3.0
